acceptably work around the bug in NavLink
NavLink's `isActive` does not substitute `%20` for ` ` the way that
`params` does

If you have a Ledger with a space in the name, you can still see this
bug manifest if you click the gear icon to edit your ledger name, then
refresh the page. The gear icon changes from black to blue again, as if
you're no longer editing this ledger, even though all the `params` know
that you still are.

The bug is unnoticeable enough now that it's not worth further
workarounds. If `react-router` fixes `NavLink` someday, maybe we can
switch back, but it's not even bad the way it is now.