move loader overload out of routes/ledgers/index
I'm not sure which of these will cause more problems down the road.

In the new approach, `routes/ledgers/new.tsx` knows about which path
it's served from, duplicating this logic from `routes/ledgers/index`.

In the old approach, redirect logic was "hidden" inside a "secret
loader" in `routes/ledgers/index`. If someone goes looking for that
logic in the future, it might take a while to find.

Another option might be to `export const path` from all of these route
files, moving this logic entirely out of `routes/ledgers/index`. That
seems hard to look at, though--a benefit of keeping the `path` in the
index is that it's easy to look at and understand how all the routes
work together, rather than hiding literally all route information in
separate files.