avoid loading Google scripts unless asked to
The old approach loaded the Google-related `<script>` tags onto the page
for every single user, even if they never clicked the "Sync with Google"
button.

This is invasive and shares too much information with Google.

This approach avoids that and leads to better UX, showing the button
immediately rather than having it flash onto the page after API scripts
finish loading.

For users who DO want Google, this also remembers that and avoids
waiting for the button click for the initial script fetching.